+++ src/styles/components/footer.scss
