+++ src/styles/components/instructions.scss
