+++ src/styles/components/loading.scss
